[11699 PART 1] Pull building queue from config out of colocated task list view
